backend/requirements.txt
fastapi==0.109.0
uvicorn
psycopg2-binary